# BuildResidentialHPXML measure: Faster stochastic schedules, second pass