--- a/BuildResidentialHPXML/resources/schedules.rb
+++ b/BuildResidentialHPXML/resources/schedules.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'csv'
-require 'yaml'
+require 'json'
 require 'matrix'
 
 class ScheduleGenerator
@@ -356,7 +356,7 @@ class ScheduleGenerator
     prng = Random.new(get_random_seed)
 
     # load the schedule configuration file
-    schedule_config = YAML.load_file(args[:resources_path] + '/schedules_config.yml')
+    schedule_config = JSON.parse(File.read(args[:resources_path] + '/schedules_config.json'))
 
     # pre-load the probability distribution csv files for speed
     cluster_size_prob_map = read_activity_cluster_size_probs(resources_path: args[:resources_path])
@@ -371,7 +371,7 @@ class ScheduleGenerator
     # if geometry_num_occupants = 2, period_in_a_year = 35040,  num_of_states = 7, then
     # shape of all_simulated_values is [2, 35040, 7]
     (1..args[:geometry_num_occupants]).each do |i|
-      occ_type_id = weighted_random(prng, schedule_config['occupancy_types_probability'])
+      occ_type_id = weighted_random(prng, schedule_config['occupancy_types']['probabilities'])
       init_prob_file_weekday = args[:resources_path] + "/schedules_weekday_mkv_chain_initial_prob_cluster_#{occ_type_id}.csv"
       initial_prob_weekday = CSV.read(init_prob_file_weekday)
       initial_prob_weekday = initial_prob_weekday.map { |x| x[0].to_f }
@@ -514,7 +514,7 @@ class ScheduleGenerator
     events_per_cluster_probs = schedule_config['sink']['events_per_cluster_probs']
     hourly_onset_prob = schedule_config['sink']['hourly_onset_prob']
     total_clusters = schedule_config['sink']['total_annual_cluster']
-    sink_between_event_gap = schedule_config['sink']['between_event_gap']
+    sink_minutes_between_event_gap = schedule_config['sink']['minutes_between_event_gap']
     cluster_per_day = total_clusters / @total_days_in_year
     sink_flow_rate_mean = schedule_config['sink']['flow_rate_mean']
     sink_flow_rate_std = schedule_config['sink']['flow_rate_std']
@@ -536,7 +536,7 @@ class ScheduleGenerator
           duration = weighted_random(prng, sink_duration_probs) + 1
           if start_min + duration > end_min then duration = (end_min - start_min) end
           sink_activity_sch.fill(sink_flow_rate, (day * 1440) + start_min, duration)
-          start_min += duration + sink_between_event_gap # Two minutes gap between sink activity
+          start_min += duration + sink_minutes_between_event_gap # Two minutes gap between sink activity
           if start_min >= end_min then break end
         end
       end
@@ -553,14 +553,14 @@ class ScheduleGenerator
     #   a. Determine the number of events in the shower cluster (there can be multiple showers)
     #   b. For each event, sample the shower duration
     #   c. Fill in the time period of personal hygiene using that many events of corresponding duration
-    #      separated by shower_between_event_gap.
+    #      separated by shower_minutes_between_event_gap.
     #      TODO If there is room in the mkc personal hygiene slot, shift uniform randomly
     # 5. If it is bath
     #   a. Sample the bath duration
     #   b. Fill in the mkc personal hygiene slot with the bath duration and flow rate.
     #      TODO If there is room in the mkc personal hygiene slot, shift uniform randomly
     # 6. Repeat process 2-6 for each occupant
-    shower_between_event_gap = schedule_config['shower']['between_event_gap']
+    shower_minutes_between_event_gap = schedule_config['shower']['minutes_between_event_gap']
     shower_flow_rate_mean = schedule_config['shower']['flow_rate_mean']
     shower_flow_rate_std = schedule_config['shower']['flow_rate_std']
     bath_ratio = schedule_config['bath']['bath_to_shower_ratio']
@@ -611,7 +611,7 @@ class ScheduleGenerator
                 m += 1
                 if (start_min + m) >= mins_in_year then break end
               end
-              shower_between_event_gap.times do
+              shower_minutes_between_event_gap.times do
                 # skip the gap between events
                 m += 1
                 if (start_min + m) >= mins_in_year then break end
@@ -634,7 +634,7 @@ class ScheduleGenerator
     # 5. Fill in the dishwasher/clothes washer time slot using those water draw events
     dw_flow_rate_mean = schedule_config['dishwasher']['flow_rate_mean']
     dw_flow_rate_std = schedule_config['dishwasher']['flow_rate_std']
-    dw_between_event_gap = schedule_config['dishwasher']['between_event_gap']
+    dw_minutes_between_event_gap = schedule_config['dishwasher']['minutes_between_event_gap']
     dw_activity_sch = [0] * mins_in_year
     m = 0
     dw_flow_rate = gaussian_rand(prng, dw_flow_rate_mean, dw_flow_rate_std, 0)
@@ -660,7 +660,7 @@ class ScheduleGenerator
           end
           if start_minute + m >= mins_in_year then break end
 
-          dw_between_event_gap.times do
+          dw_minutes_between_event_gap.times do
             m += 1
             if start_minute + m >= mins_in_year then break end
           end
@@ -673,7 +673,7 @@ class ScheduleGenerator
 
     cw_flow_rate_mean = schedule_config['clothes_washer']['flow_rate_mean']
     cw_flow_rate_std = schedule_config['clothes_washer']['flow_rate_std']
-    cw_between_event_gap = schedule_config['clothes_washer']['between_event_gap']
+    cw_minutes_between_event_gap = schedule_config['clothes_washer']['minutes_between_event_gap']
     cw_activity_sch = [0] * mins_in_year # this is the clothes_washer water draw schedule
     cw_load_size_probability = schedule_config['clothes_washer']['load_size_probability']
     m = 0
@@ -701,7 +701,7 @@ class ScheduleGenerator
             end
             if start_minute + m >= mins_in_year then break end
 
-            cw_between_event_gap.times do
+            cw_minutes_between_event_gap.times do
               # skip the gap between events
               m += 1
               if start_minute + m >= mins_in_year then break end