--- a/BuildResidentialHPXML/resources/schedules_config.json
+++ b/BuildResidentialHPXML/resources/schedules_config.json
@@ -1,7 +1,5 @@
 {
-  "_comments_": "Probabilities for all water draw events are extracted from DHW event generators. the onset, duration, events_per_cluster_probs, flow rate mean and std could all refer to the DHW event generator excel sheet ('event characteristics' and 'Start Times' sheet)",
   "occupancy_types": {
-    "_comments_": "Probability for different occupancy cluster types, Mostly Home, Early Regular Worker, Mostly Away, Regular Worker these probabilities are derived from ATUS using the k-modes algorithm",
     "probabilities": [
       0.381,
       0.297,
@@ -10,7 +8,6 @@
     ]
   },
   "plugload": {
-    "_comments_": "This is the baseline schedule for plugload, lighting and ceiling fan. It will be modified based on occupancy",
     "weekday_sch": [
       0.035,
       0.033,
@@ -79,7 +76,6 @@
     ]
   },
   "lighting": {
-    "_comments_": "The exterior and garage lighting makes use of the weekday/weekend schedule. Indoor lighting schedule is generated on the fly. Holiday lighting makes use of the holiday_sch",
     "weekday_sch": [
       0.04,
       0.037,
@@ -242,7 +238,6 @@
     ]
   },
   "cooking": {
-    "_comments_": "Monthly energy use multipliers for cooking stove/oven/range from average of multiple end-use submetering datasets (HEMS, RBSAM, ELCAP, Mass Res 1, Pecan St.)",
     "monthly_multiplier": [
       1.038,
       1.026,
@@ -259,7 +254,6 @@
     ]
   },
   "clothes_dryer": {
-    "_comments_": "Monthly energy use multipliers for clothes dryer from average of multiple end-use submetering datasets (HEMS, RBSAM, ELCAP, Mass Res 1, Pecan St., FSEC). Power draw distribution is based on csv files.",
     "monthly_multiplier": [
       1.09,
       1.054,
@@ -348,7 +342,6 @@
     "flow_rate_std": 1.17
   },
   "dishwasher": {
-    "_comments_": "Monthly energy use multipliers from average of multiple end-use submetering datasets (generally HEMS, RBSAM, ELCAP, Mass Res 1, Pecan St., and FSEC). Power draw distribution is based on csv files.",
     "flow_rate_mean": 1.39,
     "flow_rate_std": 0.2,
     "minutes_between_event_gap": 10,
@@ -368,7 +361,6 @@
     ]
   },
   "clothes_washer": {
-    "_comments_": "Monthly energy use multipliers from average of multiple end-use submetering datasets (generally HEMS, RBSAM, ELCAP, Mass Res 1, Pecan St.). Power draw distribution is based on csv files.",
     "flow_rate_mean": 2.2,
     "flow_rate_std": 0.62,
     "minutes_between_event_gap": 4,