## Pull Request Description

Some more minimal changes that reduce the ScheduleGenerator class runtime further. Now we're at ~80% reduction in speed from the original code:


  | Orig Time (s) | Prev PR Time (s) | New PR Time (s) | Tot Change
-- | -- | -- | -- | --
require yaml/json | 0.22 | 0.22 | 0.01 | -95%
ScheduleGenerator.new | 0.00 | 0.00 | 0.00 | --
ScheduleGenerator.create | 7.11 | 3.44 | 1.06 | -85%
ScheduleGenerator.export | 0.29 | 0.29 | 0.28 | --
**Sum** | **7.62** | **3.95** | **1.35** | **-82%**



Once again tested against base-schedules-stochastic.osw, where an identical base-schedules-stochastic_schedules.csv is produced before and after this change.

## Checklist

Not all may apply:

- [ ] EPvalidator.xml has been updated
- [ ] Tests (and test files) have been updated
- [ ] Documentation has been updated
- [ ] Changelog has been updated
- [ ] `openstudio tasks.rb update_measures` has been run
- [ ] No unexpected regression test changes on CI (checked comparison artifacts)