@shorowit 2021-04-12
> @shorowit Why does ScheduleGenerator.export become faster?

It doesn't become faster...?